zoos/personalities_zoo/creativity/DiagramAI/requirements.txt
Pillow
diagrams
graphviz